Rename types
